AAS 03/09/2019 11:22, Mantenimientos: Descuento.

+++ app/src/main/java/com/dtsgt/mpos/Exist.java
@@ -62,6 +62,7 @@ public class Exist extends PBase {
 		
 		super.InitBase();
 		addlog("Exist",""+du.getActDateTime(),gl.vend);
+		if(gl.reportList) addlog("Exist_Report",""+du.getActDateTime(),gl.vend);
 		
 		tipo=((appGlobals) vApp).tipo;
 
@@ -74,12 +75,14 @@ public class Exist extends PBase {
 		lblReg = (TextView) findViewById(R.id.textView1);lblReg.setText("");
 
 		setHandlers();
-		
-		try {
-			sql="DELETE FROM P_STOCK WHERE CANT+CANTM=0";
-			db.execSQL(sql);
-		} catch (SQLException e) {
-			addlog(new Object(){}.getClass().getEnclosingMethod().getName(),e.getMessage(),sql);
+
+		if (!gl.reportList) {
+			try {
+				sql="DELETE FROM P_STOCK WHERE CANT+CANTM=0";
+				db.execSQL(sql);
+			} catch (SQLException e) {
+				addlog(new Object(){}.getClass().getEnclosingMethod().getName(),e.getMessage(),sql);
+			}
 		}
 		
 		rep=new clsRepBuilder(this,gl.prw,false,gl.peMon,gl.peDecImp, "");
